Create clean_db_during_node_replacement.sh
Team, 

Since we’ve been seeing so many node failures, I put this together to check and clean the db during a node replacement. If there are remnant entries in the db tied to either the hostname or the new nodeuuid, the node will not be able to get network configurations from the IS master, and will fail to install. This script will check for and remove those entries, so the node can obtain network configs and continue with install on the next pxe boot.

This should only be used if the first attempt at replacement fails. After cleaning, please run the replace_node script again with the –r option, and proceed with pxe boot.

Let me know if you have any issues with it.